% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Census2016.DataPack-package.R
\docType{package}
\name{Census2016.DataPack-package}
\alias{Census2016.DataPack}
\alias{Census2016.DataPack-package}
\title{Census 2016 package}
\description{
Data packs from the 2016 Australian Census.
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

}
\keyword{internal}
