% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-data.R
\docType{data}
\name{Census2016__Data}
\alias{Census2016__Data}
\alias{STE__Age}
\alias{STE__Age.min}
\alias{STE__Age.min_Sex}
\alias{STE__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max}
\alias{STE__Age_HoursHousekeeping_Sex}
\alias{STE__Age_IncomeTotPersonal.max_Sex}
\alias{STE__Age_MaritalStatus_Registered}
\alias{STE__Age_MaritalStatus_Sex}
\alias{STE__Age_MaxSchoolingCompleted_Sex}
\alias{STE__Age_NeedsAssistance_Sex}
\alias{STE__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex}
\alias{STE__Age_ProvidedUnpaidDisabilityAssistance_Sex}
\alias{STE__Age_Sex}
\alias{STE__Age_Sex_Volunteer}
\alias{STE__Age_UsualResidence}
\alias{STE__Age5yr_ChildrenEverBorn}
\alias{STE__Age5yr_Indigenous}
\alias{STE__Age5yr_Indigenous_Sex}
\alias{STE__AgeChild_CoupleFamily_ParentsBornAus}
\alias{STE__AgeStudent}
\alias{STE__AgeStudent_Sex}
\alias{STE__Ancestry}
\alias{STE__Ancestry_FatherBornAus_MotherBornAus}
\alias{STE__AustCitizen}
\alias{STE__AustCitizen_Sex}
\alias{STE__BornAust}
\alias{STE__BornAust_Sex}
\alias{STE__CountryOfBirth}
\alias{STE__CountryOfBirth_Sex}
\alias{STE__CountryOfBirth_YearOfArrival.max}
\alias{STE__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min}
\alias{STE__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild}
\alias{STE__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families}
\alias{STE__CoupleFamily_HasChild_IncomeFamily.max}
\alias{STE__DwellingStructure_DwellingSubtype_HouseholdComposition}
\alias{STE__DwellingStructure_DwellingSubtype_NumberBedrooms.min}
\alias{STE__DwellingStructure_InternetAccessedFromDwelling}
\alias{STE__DwellingStructure_MortgageRepayment.min}
\alias{STE__EduInstitutionType_Sex}
\alias{STE__FamilyComposition}
\alias{STE__FamilyHousehold_IncomeHousehold.max}
\alias{STE__FamilyHousehold_UsualResidents.min}
\alias{STE__HomeCensusNight}
\alias{STE__HomeCensusNight_Sex}
\alias{STE__IndigenousStatus}
\alias{STE__Landlord_Rent.min}
\alias{STE__MaxSchoolingCompleted}
\alias{STE__MaxSchoolingCompleted_Sex}
\alias{STE__medianMortgageRepayment}
\alias{STE__medianRent}
\alias{STE__medianTotalFamilyIncome}
\alias{STE__medianTotalHouseholdIncome}
\alias{STE__medianTotalPersonalIncome}
\alias{STE__MotorVehicles.min}
\alias{STE__OnlyEnglishSpokenHome}
\alias{STE__OnlyEnglishSpokenHome_Sex}
\alias{STE__Persons}
\alias{STE__PrivateDwelling}
\alias{STE__PrivateDwelling_Sex}
\alias{STE__Religion_Denomination_Sex}
\alias{STE__Sex}
\alias{SED__Age}
\alias{SED__Age.min}
\alias{SED__Age.min_Sex}
\alias{SED__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max}
\alias{SED__Age_HoursHousekeeping_Sex}
\alias{SED__Age_IncomeTotPersonal.max_Sex}
\alias{SED__Age_MaritalStatus_Registered}
\alias{SED__Age_MaritalStatus_Sex}
\alias{SED__Age_MaxSchoolingCompleted_Sex}
\alias{SED__Age_NeedsAssistance_Sex}
\alias{SED__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex}
\alias{SED__Age_ProvidedUnpaidDisabilityAssistance_Sex}
\alias{SED__Age_Sex}
\alias{SED__Age_Sex_Volunteer}
\alias{SED__Age_UsualResidence}
\alias{SED__Age5yr_ChildrenEverBorn}
\alias{SED__Age5yr_Indigenous}
\alias{SED__Age5yr_Indigenous_Sex}
\alias{SED__AgeChild_CoupleFamily_ParentsBornAus}
\alias{SED__AgeStudent}
\alias{SED__AgeStudent_Sex}
\alias{SED__Ancestry}
\alias{SED__Ancestry_FatherBornAus_MotherBornAus}
\alias{SED__AustCitizen}
\alias{SED__AustCitizen_Sex}
\alias{SED__BornAust}
\alias{SED__BornAust_Sex}
\alias{SED__CountryOfBirth}
\alias{SED__CountryOfBirth_Sex}
\alias{SED__CountryOfBirth_YearOfArrival.max}
\alias{SED__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min}
\alias{SED__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild}
\alias{SED__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families}
\alias{SED__CoupleFamily_HasChild_IncomeFamily.max}
\alias{SED__DwellingStructure_DwellingSubtype_HouseholdComposition}
\alias{SED__DwellingStructure_DwellingSubtype_NumberBedrooms.min}
\alias{SED__DwellingStructure_InternetAccessedFromDwelling}
\alias{SED__DwellingStructure_MortgageRepayment.min}
\alias{SED__EduInstitutionType_Sex}
\alias{SED__FamilyComposition}
\alias{SED__FamilyHousehold_IncomeHousehold.max}
\alias{SED__FamilyHousehold_UsualResidents.min}
\alias{SED__HomeCensusNight}
\alias{SED__HomeCensusNight_Sex}
\alias{SED__IndigenousStatus}
\alias{SED__Landlord_Rent.min}
\alias{SED__MaxSchoolingCompleted}
\alias{SED__MaxSchoolingCompleted_Sex}
\alias{SED__medianMortgageRepayment}
\alias{SED__medianRent}
\alias{SED__medianTotalFamilyIncome}
\alias{SED__medianTotalHouseholdIncome}
\alias{SED__medianTotalPersonalIncome}
\alias{SED__MotorVehicles.min}
\alias{SED__OnlyEnglishSpokenHome}
\alias{SED__OnlyEnglishSpokenHome_Sex}
\alias{SED__Persons}
\alias{SED__PrivateDwelling}
\alias{SED__PrivateDwelling_Sex}
\alias{SED__Religion_Denomination_Sex}
\alias{SED__Sex}
\alias{SA2__Age}
\alias{SA2__Age.min}
\alias{SA2__Age.min_Sex}
\alias{SA2__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max}
\alias{SA2__Age_HoursHousekeeping_Sex}
\alias{SA2__Age_IncomeTotPersonal.max_Sex}
\alias{SA2__Age_MaritalStatus_Registered}
\alias{SA2__Age_MaritalStatus_Sex}
\alias{SA2__Age_MaxSchoolingCompleted_Sex}
\alias{SA2__Age_NeedsAssistance_Sex}
\alias{SA2__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex}
\alias{SA2__Age_ProvidedUnpaidDisabilityAssistance_Sex}
\alias{SA2__Age_Sex}
\alias{SA2__Age_Sex_Volunteer}
\alias{SA2__Age_UsualResidence}
\alias{SA2__Age5yr_ChildrenEverBorn}
\alias{SA2__Age5yr_Indigenous}
\alias{SA2__Age5yr_Indigenous_Sex}
\alias{SA2__AgeChild_CoupleFamily_ParentsBornAus}
\alias{SA2__AgeStudent}
\alias{SA2__AgeStudent_Sex}
\alias{SA2__Ancestry}
\alias{SA2__Ancestry_FatherBornAus_MotherBornAus}
\alias{SA2__AustCitizen}
\alias{SA2__AustCitizen_Sex}
\alias{SA2__BornAust}
\alias{SA2__BornAust_Sex}
\alias{SA2__CountryOfBirth}
\alias{SA2__CountryOfBirth_Sex}
\alias{SA2__CountryOfBirth_YearOfArrival.max}
\alias{SA2__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min}
\alias{SA2__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild}
\alias{SA2__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families}
\alias{SA2__CoupleFamily_HasChild_IncomeFamily.max}
\alias{SA2__DwellingStructure_DwellingSubtype_HouseholdComposition}
\alias{SA2__DwellingStructure_DwellingSubtype_NumberBedrooms.min}
\alias{SA2__DwellingStructure_InternetAccessedFromDwelling}
\alias{SA2__DwellingStructure_MortgageRepayment.min}
\alias{SA2__EduInstitutionType_Sex}
\alias{SA2__FamilyComposition}
\alias{SA2__FamilyHousehold_IncomeHousehold.max}
\alias{SA2__FamilyHousehold_UsualResidents.min}
\alias{SA2__HomeCensusNight}
\alias{SA2__HomeCensusNight_Sex}
\alias{SA2__IndigenousStatus}
\alias{SA2__Landlord_Rent.min}
\alias{SA2__MaxSchoolingCompleted}
\alias{SA2__MaxSchoolingCompleted_Sex}
\alias{SA2__medianMortgageRepayment}
\alias{SA2__medianRent}
\alias{SA2__medianTotalFamilyIncome}
\alias{SA2__medianTotalHouseholdIncome}
\alias{SA2__medianTotalPersonalIncome}
\alias{SA2__MotorVehicles.min}
\alias{SA2__OnlyEnglishSpokenHome}
\alias{SA2__OnlyEnglishSpokenHome_Sex}
\alias{SA2__Persons}
\alias{SA2__PrivateDwelling}
\alias{SA2__PrivateDwelling_Sex}
\alias{SA2__Religion_Denomination_Sex}
\alias{SA2__Sex}
\alias{SA3__Age}
\alias{SA3__Age.min}
\alias{SA3__Age.min_Sex}
\alias{SA3__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max}
\alias{SA3__Age_HoursHousekeeping_Sex}
\alias{SA3__Age_IncomeTotPersonal.max_Sex}
\alias{SA3__Age_MaritalStatus_Registered}
\alias{SA3__Age_MaritalStatus_Sex}
\alias{SA3__Age_MaxSchoolingCompleted_Sex}
\alias{SA3__Age_NeedsAssistance_Sex}
\alias{SA3__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex}
\alias{SA3__Age_ProvidedUnpaidDisabilityAssistance_Sex}
\alias{SA3__Age_Sex}
\alias{SA3__Age_Sex_Volunteer}
\alias{SA3__Age_UsualResidence}
\alias{SA3__Age5yr_ChildrenEverBorn}
\alias{SA3__Age5yr_Indigenous}
\alias{SA3__Age5yr_Indigenous_Sex}
\alias{SA3__AgeChild_CoupleFamily_ParentsBornAus}
\alias{SA3__AgeStudent}
\alias{SA3__AgeStudent_Sex}
\alias{SA3__Ancestry}
\alias{SA3__Ancestry_FatherBornAus_MotherBornAus}
\alias{SA3__AustCitizen}
\alias{SA3__AustCitizen_Sex}
\alias{SA3__BornAust}
\alias{SA3__BornAust_Sex}
\alias{SA3__CountryOfBirth}
\alias{SA3__CountryOfBirth_Sex}
\alias{SA3__CountryOfBirth_YearOfArrival.max}
\alias{SA3__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min}
\alias{SA3__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild}
\alias{SA3__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families}
\alias{SA3__CoupleFamily_HasChild_IncomeFamily.max}
\alias{SA3__DwellingStructure_DwellingSubtype_HouseholdComposition}
\alias{SA3__DwellingStructure_DwellingSubtype_NumberBedrooms.min}
\alias{SA3__DwellingStructure_InternetAccessedFromDwelling}
\alias{SA3__DwellingStructure_MortgageRepayment.min}
\alias{SA3__EduInstitutionType_Sex}
\alias{SA3__FamilyComposition}
\alias{SA3__FamilyHousehold_IncomeHousehold.max}
\alias{SA3__FamilyHousehold_UsualResidents.min}
\alias{SA3__HomeCensusNight}
\alias{SA3__HomeCensusNight_Sex}
\alias{SA3__IndigenousStatus}
\alias{SA3__Landlord_Rent.min}
\alias{SA3__MaxSchoolingCompleted}
\alias{SA3__MaxSchoolingCompleted_Sex}
\alias{SA3__medianMortgageRepayment}
\alias{SA3__medianRent}
\alias{SA3__medianTotalFamilyIncome}
\alias{SA3__medianTotalHouseholdIncome}
\alias{SA3__medianTotalPersonalIncome}
\alias{SA3__MotorVehicles.min}
\alias{SA3__OnlyEnglishSpokenHome}
\alias{SA3__OnlyEnglishSpokenHome_Sex}
\alias{SA3__Persons}
\alias{SA3__PrivateDwelling}
\alias{SA3__PrivateDwelling_Sex}
\alias{SA3__Religion_Denomination_Sex}
\alias{SA3__Sex}
\alias{SA4__Age}
\alias{SA4__Age.min}
\alias{SA4__Age.min_Sex}
\alias{SA4__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max}
\alias{SA4__Age_HoursHousekeeping_Sex}
\alias{SA4__Age_IncomeTotPersonal.max_Sex}
\alias{SA4__Age_MaritalStatus_Registered}
\alias{SA4__Age_MaritalStatus_Sex}
\alias{SA4__Age_MaxSchoolingCompleted_Sex}
\alias{SA4__Age_NeedsAssistance_Sex}
\alias{SA4__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex}
\alias{SA4__Age_ProvidedUnpaidDisabilityAssistance_Sex}
\alias{SA4__Age_Sex}
\alias{SA4__Age_Sex_Volunteer}
\alias{SA4__Age_UsualResidence}
\alias{SA4__Age5yr_ChildrenEverBorn}
\alias{SA4__Age5yr_Indigenous}
\alias{SA4__Age5yr_Indigenous_Sex}
\alias{SA4__AgeChild_CoupleFamily_ParentsBornAus}
\alias{SA4__AgeStudent}
\alias{SA4__AgeStudent_Sex}
\alias{SA4__Ancestry}
\alias{SA4__Ancestry_FatherBornAus_MotherBornAus}
\alias{SA4__AustCitizen}
\alias{SA4__AustCitizen_Sex}
\alias{SA4__BornAust}
\alias{SA4__BornAust_Sex}
\alias{SA4__CountryOfBirth}
\alias{SA4__CountryOfBirth_Sex}
\alias{SA4__CountryOfBirth_YearOfArrival.max}
\alias{SA4__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min}
\alias{SA4__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild}
\alias{SA4__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families}
\alias{SA4__CoupleFamily_HasChild_IncomeFamily.max}
\alias{SA4__DwellingStructure_DwellingSubtype_HouseholdComposition}
\alias{SA4__DwellingStructure_DwellingSubtype_NumberBedrooms.min}
\alias{SA4__DwellingStructure_InternetAccessedFromDwelling}
\alias{SA4__DwellingStructure_MortgageRepayment.min}
\alias{SA4__EduInstitutionType_Sex}
\alias{SA4__FamilyComposition}
\alias{SA4__FamilyHousehold_IncomeHousehold.max}
\alias{SA4__FamilyHousehold_UsualResidents.min}
\alias{SA4__HomeCensusNight}
\alias{SA4__HomeCensusNight_Sex}
\alias{SA4__IndigenousStatus}
\alias{SA4__Landlord_Rent.min}
\alias{SA4__MaxSchoolingCompleted}
\alias{SA4__MaxSchoolingCompleted_Sex}
\alias{SA4__medianMortgageRepayment}
\alias{SA4__medianRent}
\alias{SA4__medianTotalFamilyIncome}
\alias{SA4__medianTotalHouseholdIncome}
\alias{SA4__medianTotalPersonalIncome}
\alias{SA4__MotorVehicles.min}
\alias{SA4__OnlyEnglishSpokenHome}
\alias{SA4__OnlyEnglishSpokenHome_Sex}
\alias{SA4__Persons}
\alias{SA4__PrivateDwelling}
\alias{SA4__PrivateDwelling_Sex}
\alias{SA4__Religion_Denomination_Sex}
\alias{SA4__Sex}
\alias{LGA__Age}
\alias{LGA__Age.min}
\alias{LGA__Age.min_Sex}
\alias{LGA__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max}
\alias{LGA__Age_HoursHousekeeping_Sex}
\alias{LGA__Age_IncomeTotPersonal.max_Sex}
\alias{LGA__Age_MaritalStatus_Registered}
\alias{LGA__Age_MaritalStatus_Sex}
\alias{LGA__Age_MaxSchoolingCompleted_Sex}
\alias{LGA__Age_NeedsAssistance_Sex}
\alias{LGA__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex}
\alias{LGA__Age_ProvidedUnpaidDisabilityAssistance_Sex}
\alias{LGA__Age_Sex}
\alias{LGA__Age_Sex_Volunteer}
\alias{LGA__Age_UsualResidence}
\alias{LGA__Age5yr_ChildrenEverBorn}
\alias{LGA__Age5yr_Indigenous}
\alias{LGA__Age5yr_Indigenous_Sex}
\alias{LGA__AgeChild_CoupleFamily_ParentsBornAus}
\alias{LGA__AgeStudent}
\alias{LGA__AgeStudent_Sex}
\alias{LGA__Ancestry}
\alias{LGA__Ancestry_FatherBornAus_MotherBornAus}
\alias{LGA__AustCitizen}
\alias{LGA__AustCitizen_Sex}
\alias{LGA__BornAust}
\alias{LGA__BornAust_Sex}
\alias{LGA__CountryOfBirth}
\alias{LGA__CountryOfBirth_Sex}
\alias{LGA__CountryOfBirth_YearOfArrival.max}
\alias{LGA__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min}
\alias{LGA__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild}
\alias{LGA__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families}
\alias{LGA__CoupleFamily_HasChild_IncomeFamily.max}
\alias{LGA__DwellingStructure_DwellingSubtype_HouseholdComposition}
\alias{LGA__DwellingStructure_DwellingSubtype_NumberBedrooms.min}
\alias{LGA__DwellingStructure_InternetAccessedFromDwelling}
\alias{LGA__DwellingStructure_MortgageRepayment.min}
\alias{LGA__EduInstitutionType_Sex}
\alias{LGA__FamilyComposition}
\alias{LGA__FamilyHousehold_IncomeHousehold.max}
\alias{LGA__FamilyHousehold_UsualResidents.min}
\alias{LGA__HomeCensusNight}
\alias{LGA__HomeCensusNight_Sex}
\alias{LGA__IndigenousStatus}
\alias{LGA__Landlord_Rent.min}
\alias{LGA__MaxSchoolingCompleted}
\alias{LGA__MaxSchoolingCompleted_Sex}
\alias{LGA__MotorVehicles.min}
\alias{LGA__OnlyEnglishSpokenHome}
\alias{LGA__OnlyEnglishSpokenHome_Sex}
\alias{LGA__Persons}
\alias{LGA__PrivateDwelling}
\alias{LGA__PrivateDwelling_Sex}
\alias{LGA__Religion_Denomination_Sex}
\alias{LGA__Sex}
\alias{CED__Age}
\alias{CED__Age.min}
\alias{CED__Age.min_Sex}
\alias{CED__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max}
\alias{CED__Age_HoursHousekeeping_Sex}
\alias{CED__Age_IncomeTotPersonal.max_Sex}
\alias{CED__Age_MaritalStatus_Registered}
\alias{CED__Age_MaritalStatus_Sex}
\alias{CED__Age_MaxSchoolingCompleted_Sex}
\alias{CED__Age_NeedsAssistance_Sex}
\alias{CED__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex}
\alias{CED__Age_ProvidedUnpaidDisabilityAssistance_Sex}
\alias{CED__Age_Sex}
\alias{CED__Age_Sex_Volunteer}
\alias{CED__Age_UsualResidence}
\alias{CED__Age5yr_ChildrenEverBorn}
\alias{CED__Age5yr_Indigenous}
\alias{CED__Age5yr_Indigenous_Sex}
\alias{CED__AgeChild_CoupleFamily_ParentsBornAus}
\alias{CED__AgeStudent}
\alias{CED__AgeStudent_Sex}
\alias{CED__Ancestry}
\alias{CED__Ancestry_FatherBornAus_MotherBornAus}
\alias{CED__AustCitizen}
\alias{CED__AustCitizen_Sex}
\alias{CED__BornAust}
\alias{CED__BornAust_Sex}
\alias{CED__CountryOfBirth}
\alias{CED__CountryOfBirth_Sex}
\alias{CED__CountryOfBirth_YearOfArrival.max}
\alias{CED__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min}
\alias{CED__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild}
\alias{CED__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families}
\alias{CED__CoupleFamily_HasChild_IncomeFamily.max}
\alias{CED__DwellingStructure_DwellingSubtype_HouseholdComposition}
\alias{CED__DwellingStructure_DwellingSubtype_NumberBedrooms.min}
\alias{CED__DwellingStructure_InternetAccessedFromDwelling}
\alias{CED__DwellingStructure_MortgageRepayment.min}
\alias{CED__EduInstitutionType_Sex}
\alias{CED__FamilyComposition}
\alias{CED__FamilyHousehold_IncomeHousehold.max}
\alias{CED__FamilyHousehold_UsualResidents.min}
\alias{CED__HomeCensusNight}
\alias{CED__HomeCensusNight_Sex}
\alias{CED__IndigenousStatus}
\alias{CED__Landlord_Rent.min}
\alias{CED__MaxSchoolingCompleted}
\alias{CED__MaxSchoolingCompleted_Sex}
\alias{CED__MotorVehicles.min}
\alias{CED__OnlyEnglishSpokenHome}
\alias{CED__OnlyEnglishSpokenHome_Sex}
\alias{CED__Persons}
\alias{CED__PrivateDwelling}
\alias{CED__PrivateDwelling_Sex}
\alias{CED__Religion_Denomination_Sex}
\alias{CED__Sex}
\alias{POA__Age}
\alias{POA__Age.min}
\alias{POA__Age.min_Sex}
\alias{POA__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max}
\alias{POA__Age_HoursHousekeeping_Sex}
\alias{POA__Age_IncomeTotPersonal.max_Sex}
\alias{POA__Age_MaritalStatus_Registered}
\alias{POA__Age_MaritalStatus_Sex}
\alias{POA__Age_MaxSchoolingCompleted_Sex}
\alias{POA__Age_NeedsAssistance_Sex}
\alias{POA__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex}
\alias{POA__Age_ProvidedUnpaidDisabilityAssistance_Sex}
\alias{POA__Age_Sex}
\alias{POA__Age_Sex_Volunteer}
\alias{POA__Age_UsualResidence}
\alias{POA__Age5yr_ChildrenEverBorn}
\alias{POA__Age5yr_Indigenous}
\alias{POA__Age5yr_Indigenous_Sex}
\alias{POA__AgeChild_CoupleFamily_ParentsBornAus}
\alias{POA__AgeStudent}
\alias{POA__AgeStudent_Sex}
\alias{POA__Ancestry}
\alias{POA__Ancestry_FatherBornAus_MotherBornAus}
\alias{POA__AustCitizen}
\alias{POA__AustCitizen_Sex}
\alias{POA__BornAust}
\alias{POA__BornAust_Sex}
\alias{POA__CountryOfBirth}
\alias{POA__CountryOfBirth_Sex}
\alias{POA__CountryOfBirth_YearOfArrival.max}
\alias{POA__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min}
\alias{POA__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild}
\alias{POA__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families}
\alias{POA__CoupleFamily_HasChild_IncomeFamily.max}
\alias{POA__DwellingStructure_DwellingSubtype_HouseholdComposition}
\alias{POA__DwellingStructure_DwellingSubtype_NumberBedrooms.min}
\alias{POA__DwellingStructure_InternetAccessedFromDwelling}
\alias{POA__DwellingStructure_MortgageRepayment.min}
\alias{POA__EduInstitutionType_Sex}
\alias{POA__FamilyComposition}
\alias{POA__FamilyHousehold_IncomeHousehold.max}
\alias{POA__FamilyHousehold_UsualResidents.min}
\alias{POA__HomeCensusNight}
\alias{POA__HomeCensusNight_Sex}
\alias{POA__IndigenousStatus}
\alias{POA__Landlord_Rent.min}
\alias{POA__MaxSchoolingCompleted}
\alias{POA__MaxSchoolingCompleted_Sex}
\alias{POA__medianMortgageRepayment}
\alias{POA__medianRent}
\alias{POA__medianTotalFamilyIncome}
\alias{POA__medianTotalHouseholdIncome}
\alias{POA__medianTotalPersonalIncome}
\alias{POA__MotorVehicles.min}
\alias{POA__OnlyEnglishSpokenHome}
\alias{POA__OnlyEnglishSpokenHome_Sex}
\alias{POA__Persons}
\alias{POA__PrivateDwelling}
\alias{POA__PrivateDwelling_Sex}
\alias{POA__Religion_Denomination_Sex}
\alias{POA__Sex}
\title{Data pack data}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 99 rows and 3 columns.}
\usage{
STE__Age

STE__Age.min

STE__Age.min_Sex

STE__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max

STE__Age_HoursHousekeeping_Sex

STE__Age_IncomeTotPersonal.max_Sex

STE__Age_MaritalStatus_Registered

STE__Age_MaritalStatus_Sex

STE__Age_MaxSchoolingCompleted_Sex

STE__Age_NeedsAssistance_Sex

STE__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex

STE__Age_ProvidedUnpaidDisabilityAssistance_Sex

STE__Age_Sex

STE__Age_Sex_Volunteer

STE__Age_UsualResidence

STE__Age5yr_ChildrenEverBorn

STE__Age5yr_Indigenous

STE__Age5yr_Indigenous_Sex

STE__AgeChild_CoupleFamily_ParentsBornAus

STE__AgeStudent

STE__AgeStudent_Sex

STE__Ancestry

STE__Ancestry_FatherBornAus_MotherBornAus

STE__AustCitizen

STE__AustCitizen_Sex

STE__BornAust

STE__BornAust_Sex

STE__CountryOfBirth

STE__CountryOfBirth_Sex

STE__CountryOfBirth_YearOfArrival.max

STE__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min

STE__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild


  STE__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families

STE__CoupleFamily_HasChild_IncomeFamily.max

STE__DwellingStructure_DwellingSubtype_HouseholdComposition

STE__DwellingStructure_DwellingSubtype_NumberBedrooms.min

STE__DwellingStructure_InternetAccessedFromDwelling

STE__DwellingStructure_MortgageRepayment.min

STE__EduInstitutionType_Sex

STE__FamilyComposition

STE__FamilyHousehold_IncomeHousehold.max

STE__FamilyHousehold_UsualResidents.min

STE__HomeCensusNight

STE__HomeCensusNight_Sex

STE__IndigenousStatus

STE__Landlord_Rent.min

STE__MaxSchoolingCompleted

STE__MaxSchoolingCompleted_Sex

STE__medianMortgageRepayment

STE__medianRent

STE__medianTotalFamilyIncome

STE__medianTotalHouseholdIncome

STE__medianTotalPersonalIncome

STE__MotorVehicles.min

STE__OnlyEnglishSpokenHome

STE__OnlyEnglishSpokenHome_Sex

STE__Persons

STE__PrivateDwelling

STE__PrivateDwelling_Sex

STE__Religion_Denomination_Sex

STE__Sex

SED__Age

SED__Age.min

SED__Age.min_Sex

SED__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max

SED__Age_HoursHousekeeping_Sex

SED__Age_IncomeTotPersonal.max_Sex

SED__Age_MaritalStatus_Registered

SED__Age_MaritalStatus_Sex

SED__Age_MaxSchoolingCompleted_Sex

SED__Age_NeedsAssistance_Sex

SED__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex

SED__Age_ProvidedUnpaidDisabilityAssistance_Sex

SED__Age_Sex

SED__Age_Sex_Volunteer

SED__Age_UsualResidence

SED__Age5yr_ChildrenEverBorn

SED__Age5yr_Indigenous

SED__Age5yr_Indigenous_Sex

SED__AgeChild_CoupleFamily_ParentsBornAus

SED__AgeStudent

SED__AgeStudent_Sex

SED__Ancestry

SED__Ancestry_FatherBornAus_MotherBornAus

SED__AustCitizen

SED__AustCitizen_Sex

SED__BornAust

SED__BornAust_Sex

SED__CountryOfBirth

SED__CountryOfBirth_Sex

SED__CountryOfBirth_YearOfArrival.max

SED__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min

SED__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild


  SED__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families

SED__CoupleFamily_HasChild_IncomeFamily.max

SED__DwellingStructure_DwellingSubtype_HouseholdComposition

SED__DwellingStructure_DwellingSubtype_NumberBedrooms.min

SED__DwellingStructure_InternetAccessedFromDwelling

SED__DwellingStructure_MortgageRepayment.min

SED__EduInstitutionType_Sex

SED__FamilyComposition

SED__FamilyHousehold_IncomeHousehold.max

SED__FamilyHousehold_UsualResidents.min

SED__HomeCensusNight

SED__HomeCensusNight_Sex

SED__IndigenousStatus

SED__Landlord_Rent.min

SED__MaxSchoolingCompleted

SED__MaxSchoolingCompleted_Sex

SED__medianMortgageRepayment

SED__medianRent

SED__medianTotalFamilyIncome

SED__medianTotalHouseholdIncome

SED__medianTotalPersonalIncome

SED__MotorVehicles.min

SED__OnlyEnglishSpokenHome

SED__OnlyEnglishSpokenHome_Sex

SED__Persons

SED__PrivateDwelling

SED__PrivateDwelling_Sex

SED__Religion_Denomination_Sex

SED__Sex

SA2__Age

SA2__Age.min

SA2__Age.min_Sex

SA2__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max

SA2__Age_HoursHousekeeping_Sex

SA2__Age_IncomeTotPersonal.max_Sex

SA2__Age_MaritalStatus_Registered

SA2__Age_MaritalStatus_Sex

SA2__Age_MaxSchoolingCompleted_Sex

SA2__Age_NeedsAssistance_Sex

SA2__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex

SA2__Age_ProvidedUnpaidDisabilityAssistance_Sex

SA2__Age_Sex

SA2__Age_Sex_Volunteer

SA2__Age_UsualResidence

SA2__Age5yr_ChildrenEverBorn

SA2__Age5yr_Indigenous

SA2__Age5yr_Indigenous_Sex

SA2__AgeChild_CoupleFamily_ParentsBornAus

SA2__AgeStudent

SA2__AgeStudent_Sex

SA2__Ancestry

SA2__Ancestry_FatherBornAus_MotherBornAus

SA2__AustCitizen

SA2__AustCitizen_Sex

SA2__BornAust

SA2__BornAust_Sex

SA2__CountryOfBirth

SA2__CountryOfBirth_Sex

SA2__CountryOfBirth_YearOfArrival.max

SA2__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min

SA2__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild


  SA2__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families

SA2__CoupleFamily_HasChild_IncomeFamily.max

SA2__DwellingStructure_DwellingSubtype_HouseholdComposition

SA2__DwellingStructure_DwellingSubtype_NumberBedrooms.min

SA2__DwellingStructure_InternetAccessedFromDwelling

SA2__DwellingStructure_MortgageRepayment.min

SA2__EduInstitutionType_Sex

SA2__FamilyComposition

SA2__FamilyHousehold_IncomeHousehold.max

SA2__FamilyHousehold_UsualResidents.min

SA2__HomeCensusNight

SA2__HomeCensusNight_Sex

SA2__IndigenousStatus

SA2__Landlord_Rent.min

SA2__MaxSchoolingCompleted

SA2__MaxSchoolingCompleted_Sex

SA2__medianMortgageRepayment

SA2__medianRent

SA2__medianTotalFamilyIncome

SA2__medianTotalHouseholdIncome

SA2__medianTotalPersonalIncome

SA2__MotorVehicles.min

SA2__OnlyEnglishSpokenHome

SA2__OnlyEnglishSpokenHome_Sex

SA2__Persons

SA2__PrivateDwelling

SA2__PrivateDwelling_Sex

SA2__Religion_Denomination_Sex

SA2__Sex

SA3__Age

SA3__Age.min

SA3__Age.min_Sex

SA3__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max

SA3__Age_HoursHousekeeping_Sex

SA3__Age_IncomeTotPersonal.max_Sex

SA3__Age_MaritalStatus_Registered

SA3__Age_MaritalStatus_Sex

SA3__Age_MaxSchoolingCompleted_Sex

SA3__Age_NeedsAssistance_Sex

SA3__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex

SA3__Age_ProvidedUnpaidDisabilityAssistance_Sex

SA3__Age_Sex

SA3__Age_Sex_Volunteer

SA3__Age_UsualResidence

SA3__Age5yr_ChildrenEverBorn

SA3__Age5yr_Indigenous

SA3__Age5yr_Indigenous_Sex

SA3__AgeChild_CoupleFamily_ParentsBornAus

SA3__AgeStudent

SA3__AgeStudent_Sex

SA3__Ancestry

SA3__Ancestry_FatherBornAus_MotherBornAus

SA3__AustCitizen

SA3__AustCitizen_Sex

SA3__BornAust

SA3__BornAust_Sex

SA3__CountryOfBirth

SA3__CountryOfBirth_Sex

SA3__CountryOfBirth_YearOfArrival.max

SA3__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min

SA3__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild


  SA3__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families

SA3__CoupleFamily_HasChild_IncomeFamily.max

SA3__DwellingStructure_DwellingSubtype_HouseholdComposition

SA3__DwellingStructure_DwellingSubtype_NumberBedrooms.min

SA3__DwellingStructure_InternetAccessedFromDwelling

SA3__DwellingStructure_MortgageRepayment.min

SA3__EduInstitutionType_Sex

SA3__FamilyComposition

SA3__FamilyHousehold_IncomeHousehold.max

SA3__FamilyHousehold_UsualResidents.min

SA3__HomeCensusNight

SA3__HomeCensusNight_Sex

SA3__IndigenousStatus

SA3__Landlord_Rent.min

SA3__MaxSchoolingCompleted

SA3__MaxSchoolingCompleted_Sex

SA3__medianMortgageRepayment

SA3__medianRent

SA3__medianTotalFamilyIncome

SA3__medianTotalHouseholdIncome

SA3__medianTotalPersonalIncome

SA3__MotorVehicles.min

SA3__OnlyEnglishSpokenHome

SA3__OnlyEnglishSpokenHome_Sex

SA3__Persons

SA3__PrivateDwelling

SA3__PrivateDwelling_Sex

SA3__Religion_Denomination_Sex

SA3__Sex

SA4__Age

SA4__Age.min

SA4__Age.min_Sex

SA4__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max

SA4__Age_HoursHousekeeping_Sex

SA4__Age_IncomeTotPersonal.max_Sex

SA4__Age_MaritalStatus_Registered

SA4__Age_MaritalStatus_Sex

SA4__Age_MaxSchoolingCompleted_Sex

SA4__Age_NeedsAssistance_Sex

SA4__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex

SA4__Age_ProvidedUnpaidDisabilityAssistance_Sex

SA4__Age_Sex

SA4__Age_Sex_Volunteer

SA4__Age_UsualResidence

SA4__Age5yr_ChildrenEverBorn

SA4__Age5yr_Indigenous

SA4__Age5yr_Indigenous_Sex

SA4__AgeChild_CoupleFamily_ParentsBornAus

SA4__AgeStudent

SA4__AgeStudent_Sex

SA4__Ancestry

SA4__Ancestry_FatherBornAus_MotherBornAus

SA4__AustCitizen

SA4__AustCitizen_Sex

SA4__BornAust

SA4__BornAust_Sex

SA4__CountryOfBirth

SA4__CountryOfBirth_Sex

SA4__CountryOfBirth_YearOfArrival.max

SA4__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min

SA4__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild


  SA4__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families

SA4__CoupleFamily_HasChild_IncomeFamily.max

SA4__DwellingStructure_DwellingSubtype_HouseholdComposition

SA4__DwellingStructure_DwellingSubtype_NumberBedrooms.min

SA4__DwellingStructure_InternetAccessedFromDwelling

SA4__DwellingStructure_MortgageRepayment.min

SA4__EduInstitutionType_Sex

SA4__FamilyComposition

SA4__FamilyHousehold_IncomeHousehold.max

SA4__FamilyHousehold_UsualResidents.min

SA4__HomeCensusNight

SA4__HomeCensusNight_Sex

SA4__IndigenousStatus

SA4__Landlord_Rent.min

SA4__MaxSchoolingCompleted

SA4__MaxSchoolingCompleted_Sex

SA4__medianMortgageRepayment

SA4__medianRent

SA4__medianTotalFamilyIncome

SA4__medianTotalHouseholdIncome

SA4__medianTotalPersonalIncome

SA4__MotorVehicles.min

SA4__OnlyEnglishSpokenHome

SA4__OnlyEnglishSpokenHome_Sex

SA4__Persons

SA4__PrivateDwelling

SA4__PrivateDwelling_Sex

SA4__Religion_Denomination_Sex

SA4__Sex

LGA__Age

LGA__Age.min

LGA__Age.min_Sex

LGA__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max

LGA__Age_HoursHousekeeping_Sex

LGA__Age_IncomeTotPersonal.max_Sex

LGA__Age_MaritalStatus_Registered

LGA__Age_MaritalStatus_Sex

LGA__Age_MaxSchoolingCompleted_Sex

LGA__Age_NeedsAssistance_Sex

LGA__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex

LGA__Age_ProvidedUnpaidDisabilityAssistance_Sex

LGA__Age_Sex

LGA__Age_Sex_Volunteer

LGA__Age_UsualResidence

LGA__Age5yr_ChildrenEverBorn

LGA__Age5yr_Indigenous

LGA__Age5yr_Indigenous_Sex

LGA__AgeChild_CoupleFamily_ParentsBornAus

LGA__AgeStudent

LGA__AgeStudent_Sex

LGA__Ancestry

LGA__Ancestry_FatherBornAus_MotherBornAus

LGA__AustCitizen

LGA__AustCitizen_Sex

LGA__BornAust

LGA__BornAust_Sex

LGA__CountryOfBirth

LGA__CountryOfBirth_Sex

LGA__CountryOfBirth_YearOfArrival.max

LGA__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min

LGA__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild


  LGA__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families

LGA__CoupleFamily_HasChild_IncomeFamily.max

LGA__DwellingStructure_DwellingSubtype_HouseholdComposition

LGA__DwellingStructure_DwellingSubtype_NumberBedrooms.min

LGA__DwellingStructure_InternetAccessedFromDwelling

LGA__DwellingStructure_MortgageRepayment.min

LGA__EduInstitutionType_Sex

LGA__FamilyComposition

LGA__FamilyHousehold_IncomeHousehold.max

LGA__FamilyHousehold_UsualResidents.min

LGA__HomeCensusNight

LGA__HomeCensusNight_Sex

LGA__IndigenousStatus

LGA__Landlord_Rent.min

LGA__MaxSchoolingCompleted

LGA__MaxSchoolingCompleted_Sex

LGA__MotorVehicles.min

LGA__OnlyEnglishSpokenHome

LGA__OnlyEnglishSpokenHome_Sex

LGA__Persons

LGA__PrivateDwelling

LGA__PrivateDwelling_Sex

LGA__Religion_Denomination_Sex

LGA__Sex

CED__Age

CED__Age.min

CED__Age.min_Sex

CED__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max

CED__Age_HoursHousekeeping_Sex

CED__Age_IncomeTotPersonal.max_Sex

CED__Age_MaritalStatus_Registered

CED__Age_MaritalStatus_Sex

CED__Age_MaxSchoolingCompleted_Sex

CED__Age_NeedsAssistance_Sex

CED__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex

CED__Age_ProvidedUnpaidDisabilityAssistance_Sex

CED__Age_Sex

CED__Age_Sex_Volunteer

CED__Age_UsualResidence

CED__Age5yr_ChildrenEverBorn

CED__Age5yr_Indigenous

CED__Age5yr_Indigenous_Sex

CED__AgeChild_CoupleFamily_ParentsBornAus

CED__AgeStudent

CED__AgeStudent_Sex

CED__Ancestry

CED__Ancestry_FatherBornAus_MotherBornAus

CED__AustCitizen

CED__AustCitizen_Sex

CED__BornAust

CED__BornAust_Sex

CED__CountryOfBirth

CED__CountryOfBirth_Sex

CED__CountryOfBirth_YearOfArrival.max

CED__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min

CED__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild


  CED__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families

CED__CoupleFamily_HasChild_IncomeFamily.max

CED__DwellingStructure_DwellingSubtype_HouseholdComposition

CED__DwellingStructure_DwellingSubtype_NumberBedrooms.min

CED__DwellingStructure_InternetAccessedFromDwelling

CED__DwellingStructure_MortgageRepayment.min

CED__EduInstitutionType_Sex

CED__FamilyComposition

CED__FamilyHousehold_IncomeHousehold.max

CED__FamilyHousehold_UsualResidents.min

CED__HomeCensusNight

CED__HomeCensusNight_Sex

CED__IndigenousStatus

CED__Landlord_Rent.min

CED__MaxSchoolingCompleted

CED__MaxSchoolingCompleted_Sex

CED__MotorVehicles.min

CED__OnlyEnglishSpokenHome

CED__OnlyEnglishSpokenHome_Sex

CED__Persons

CED__PrivateDwelling

CED__PrivateDwelling_Sex

CED__Religion_Denomination_Sex

CED__Sex

POA__Age

POA__Age.min

POA__Age.min_Sex

POA__Age_EnglishProficiency_SpeaksEnglishOnly_YearOfArrival.max

POA__Age_HoursHousekeeping_Sex

POA__Age_IncomeTotPersonal.max_Sex

POA__Age_MaritalStatus_Registered

POA__Age_MaritalStatus_Sex

POA__Age_MaxSchoolingCompleted_Sex

POA__Age_NeedsAssistance_Sex

POA__Age_ProvidedUnpaidChildcare_ForOwnChild_ForOtherChild_Sex

POA__Age_ProvidedUnpaidDisabilityAssistance_Sex

POA__Age_Sex

POA__Age_Sex_Volunteer

POA__Age_UsualResidence

POA__Age5yr_ChildrenEverBorn

POA__Age5yr_Indigenous

POA__Age5yr_Indigenous_Sex

POA__AgeChild_CoupleFamily_ParentsBornAus

POA__AgeStudent

POA__AgeStudent_Sex

POA__Ancestry

POA__Ancestry_FatherBornAus_MotherBornAus

POA__AustCitizen

POA__AustCitizen_Sex

POA__BornAust

POA__BornAust_Sex

POA__CountryOfBirth

POA__CountryOfBirth_Sex

POA__CountryOfBirth_YearOfArrival.max

POA__CoupleFamily_HasChild_HasChildUnder15_MortgageRepayment.min

POA__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild


  POA__CoupleFamily_HasChild_HasDependentStudent_HasNonDependentChild_families

POA__CoupleFamily_HasChild_IncomeFamily.max

POA__DwellingStructure_DwellingSubtype_HouseholdComposition

POA__DwellingStructure_DwellingSubtype_NumberBedrooms.min

POA__DwellingStructure_InternetAccessedFromDwelling

POA__DwellingStructure_MortgageRepayment.min

POA__EduInstitutionType_Sex

POA__FamilyComposition

POA__FamilyHousehold_IncomeHousehold.max

POA__FamilyHousehold_UsualResidents.min

POA__HomeCensusNight

POA__HomeCensusNight_Sex

POA__IndigenousStatus

POA__Landlord_Rent.min

POA__MaxSchoolingCompleted

POA__MaxSchoolingCompleted_Sex

POA__medianMortgageRepayment

POA__medianRent

POA__medianTotalFamilyIncome

POA__medianTotalHouseholdIncome

POA__medianTotalPersonalIncome

POA__MotorVehicles.min

POA__OnlyEnglishSpokenHome

POA__OnlyEnglishSpokenHome_Sex

POA__Persons

POA__PrivateDwelling

POA__PrivateDwelling_Sex

POA__Religion_Denomination_Sex

POA__Sex
}
\description{
Data from the 2016 Australian Census
}
\keyword{datasets}
