% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ISO3166.R
\docType{data}
\name{ISO3166}
\alias{ISO3166}
\title{ISO3166 decoder}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 249 rows and 11 columns.}
\usage{
ISO3166
}
\description{
Countries with the alpha-2 and alpha-3 codes.
}
\keyword{datasets}
