% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA16_to_17.R
\name{LGA16_to_17}
\alias{LGA16_to_17}
\title{Translate LGA 2016 values to LGA 207}
\usage{
LGA16_to_17(x, method = c("sum", "average"))
}
\arguments{
\item{x}{A data frame with 2 columns, one of which must by LGA_NAME16 and the other should be some
kind of numeric value}

\item{method}{Whether the values should be combined by adding them or by taking an average weighted
by the populations of Botany Bay and Rockdale (note that these are not always the appropriate weights,
hence this method is only going to be an approximation for values that cannot be summed)}
}
\description{
Translate LGA 2016 values to LGA 207
}
\details{
In late 2016, Botany Bay (C) and Rockdale (C) merged into Bayside (A), in New South Wales.
Do not confuse this with Bayside (C), in Victoria... The Census datapack uses Botany Bay and Rockdale.
This function gives a crude approximation for Bayside by combining the rows for Botany Bay and Rockdale.
It only works with a one dimensional dataset (one row per LGA), although it would be a great extension
to let it work with arbitrary numbers of additional dimensions as well (eg age by sex at the same time).
}
\author{
Peter Ellis peter.ellis2013nz@gmail.com
}
