% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Metadata.R
\docType{data}
\name{Metadata}
\alias{Metadata}
\title{Metadata for the files}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 8616 rows and 6 columns.}
\usage{
Metadata
}
\description{
Metadata file from the ABS data packs, standardized and normalized.
}
\keyword{datasets}
