% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-NonABS.R
\docType{data}
\name{NonABS_decoder}
\alias{NonABS_decoder}
\title{NonABS_decoder}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 19153 rows and 5 columns.}
\usage{
NonABS_decoder
}
\description{
Decoder for the non-ABS geographic data structures, includign the area.
}
\details{
Variable names may change without notice.
}
\keyword{datasets}
