% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region__medianMortgageRepayment.R
\docType{data}
\name{Region__medianMortgageRepayment}
\alias{Region__medianMortgageRepayment}
\alias{LGA__medianMortgageRepayment}
\alias{CED__medianMortgageRepayment}
\title{Median mortgage repayment}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 563 rows and 2 columns.}
\usage{
LGA__medianMortgageRepayment

CED__medianMortgageRepayment
}
\details{
\describe{
\item{\code{medianMortgageRepayment}}{The \strong{annual} median mortgage repayment.
Applicable to occupied private dwellings being purchased and includes dwellings being
purchased under a shared equity scheme. It excludes 'Visitors only' and 'Other non-classifiable households'.}
}
}
\keyword{datasets}
