% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region__medianRent.R
\docType{data}
\name{Region__medianRent}
\alias{Region__medianRent}
\alias{LGA__medianRent}
\alias{CED__medianRent}
\title{Median rent}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 563 rows and 2 columns.}
\usage{
LGA__medianRent

CED__medianRent
}
\description{
Median rent
}
\details{
\describe{
\item{\code{medianRent}}{The \strong{annual} rent.
Median rent is applicable to occupied private dwellings being rented. It excludes 'Visitors only' and 'Other non-classifiable' households.
}
}
}
\keyword{datasets}
