% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region__medianTotalFamilyIncome.R
\docType{data}
\name{Region__medianTotalFamilyIncome}
\alias{Region__medianTotalFamilyIncome}
\alias{LGA__medianTotalFamilyIncome}
\alias{CED__medianTotalFamilyIncome}
\title{Median total family income}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 563 rows and 2 columns.}
\usage{
LGA__medianTotalFamilyIncome

CED__medianTotalFamilyIncome
}
\description{
Median total family income
}
\details{
\describe{
\item{\code{medianTotalPersonalIncome}}{The \strong{annual} income.
Median total household income is applicable to occupied private dwellings. It excludes households where at least one member aged 15 years and over did not state an income and households
where at least one member aged 15 years and over was temporarily absent on Census Night. It excludes 'Visitors only' and 'Other non-classifiable' households.
}
}
}
\keyword{datasets}
