% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region__medianTotalPersonalIncome.R
\docType{data}
\name{Region__medianTotalPersonalIncome}
\alias{Region__medianTotalPersonalIncome}
\alias{LGA__medianTotalPersonalIncome}
\alias{CED__medianTotalPersonalIncome}
\title{Median total personal income}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 563 rows and 2 columns.}
\usage{
LGA__medianTotalPersonalIncome

CED__medianTotalPersonalIncome
}
\description{
Median total personal income
}
\details{
\describe{
\item{\code{medianTotalPersonalIncome}}{The \strong{annual} income. Applicable to persons aged 15 years and over.}
}
}
\keyword{datasets}
