% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SA16_decoder.R
\docType{data}
\name{SA16_decoder}
\alias{SA16_decoder}
\title{SA16 decoder}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 2785 rows and 3 columns.}
\usage{
SA16_decoder
}
\description{
Decode between Census Codes and Census Names.
}
\keyword{datasets}
