% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGA16_to_17.R
\name{LGA16_to_17}
\alias{LGA16_to_17}
\title{Approximately aggregate LGA 2016 values to match LGA 2017 names}
\usage{
LGA16_to_17(x, method = c("sum", "average"),
  value_var = names(x)[ncol(x)], return. = NULL)
}
\arguments{
\item{x}{A data frame with 2 or more columns, one of which must by LGA_NAME16 and another should be some
kind of numeric value; all other columns must be appropriate for a \code{dplyr::group_by} operation
(e.g. categorical variables like age, sex and indigenous status)}

\item{method}{Whether the values should be combined by adding them (eg for head counts) or by taking
an average (e.g. for median income) weighted by the populations of Botany Bay and Rockdale (note that
these are not always the appropriate weights,
hence this method is only going to be an approximation for values that cannot be summed)}

\item{value_var}{the value variable that is to be grouped up and summarised. If not provided, this
is assumed to be the final column of x.}

\item{return.}{The class of the returned value. If \code{NULL}, the default,
a \code{data.table} is returned if \code{x} is a \code{data.table} and a \code{tibble}
is \code{x} is a \code{tibble}.}
}
\value{
A table one row shorter than \code{x} (for each group).
}
\description{
Approximately aggregate LGA 2016 values to match LGA 2017 names
}
\details{
In late 2016, Botany Bay (C) and Rockdale (C) merged into Bayside (A), in New South Wales.
Do not confuse this with Bayside (C), in Victoria... The Census datapack uses Botany Bay and Rockdale.
This function gives a crude approximation for Bayside by combining the rows for Botany Bay and Rockdale.
It only works with a one dimensional dataset (one row per LGA), although it would be a great extension
to let it work with arbitrary numbers of additional dimensions as well (eg age by sex at the same time).

Kalamunda (S) was also renamed Kalamunda (C).
}
\section{Warning}{

This is only a crude approximation of the changes to LGAs in 2017.  See
\href{http://www.abs.gov.au/ausstats/abs@.nsf/Latestproducts/3218.0Appendix12016-17?opendocument&tabname=Notes&prodno=3218.0&issue=2016-17&num=&view=}{the ABS on changes to LGAs in ASGS 2017}
for the real rundown.  In particular, no account is taken by \code{LGA16_to_17} of the shifts between
local government areas in Western Australia (eg Bayswater (C) gained from Swan (C), etc). The aim of this
function is to provide a work around for people using LGA 2017 boundaries who want at least
some approximation of census data.
}

\examples{
dim(LGA__medianTotalHouseholdIncome)
 # one row less:
dim(LGA16_to_17(LGA__medianTotalHouseholdIncome, method = "average"))

LGA17__Age_Sex <- LGA16_to_17(LGA__Age_Sex)

# 22 rows less (11 age groups and 2 sexes):
dim(LGA__Age_Sex)
dim(LGA17__Age_Sex)

# same total number of persons:
c(sum(LGA__Age_Sex$persons), sum(LGA17__Age_Sex$persons))

# check which names have appeared and disappeared
r17 <- unique(LGA__Age_Sex$LGA_NAME16)
r16 <- unique(LGA17__Age_Sex$LGA_NAME17)
r17[!r17 \%in\% r16]
r16[!r16 \%in\% r17]

}
\author{
Peter Ellis
}
